Clean up the documentation.
I had the opportunity to use this library this week after having been away
for a while, so I had a chance to see the documentation with fresh eyes.
This should make it more clear how to use the code. It also removes
needless godoc complexity by privatizing the Allocator implementations,
which are just useless godoc noise. This is a breaking change on the
off chance you directly used a GCAllocator in your code.

On the incredible long shot that anyone, anywhere, did that, replace
your

    var gcAllocator gomempool.GCAllocator

with

    var pool *gomempool.Pool
    gcAllocator := pool.GetNewAllocator()